fix: move secureRoute into component
